function/requirements.txt
google-cloud-storage
google-auth
requests
cryptography
